[zend-openid] fix for file storage on windows
fix for `Zend_OpenId_Consumer_Storage_File` when symlinks are not used (i.e. on windows)
One of the files created in `Zend_OpenId_Consumer_Storage_File::addAssociations()` was being deleted for some unknown reason, but `Zend_OpenId_Consumer` relies on it to be there.

Fixes errors in OpenId tests when running on windows. i.a.
```
Zend_Auth_Adapter_OpenIdTest::testAuthenticateLoginValid
Failed asserting that false is true.
```